Refactor to remove import loop between gmd_node <-> gmd_mesh
