+++ yk_gmd_blender/yk_gmd/v2/abstract/gmd_mesh.py
@@ -5,8 +5,8 @@ from typing import List, Tuple
 from mathutils import Vector
 
 from yk_gmd_blender.yk_gmd.v2.abstract.gmd_attributes import GMDAttributeSet
-from yk_gmd_blender.yk_gmd.v2.abstract.gmd_node import GMDBone
 from yk_gmd_blender.yk_gmd.v2.abstract.gmd_shader import GMDVertexBuffer
+from yk_gmd_blender.yk_gmd.v2.abstract.nodes.gmd_bone import GMDBone
 
 
 @dataclass
