+++ yk_gmd_blender/yk_gmd/v2/abstract/gmd_scene.py
@@ -2,7 +2,9 @@ from dataclasses import dataclass
 from queue import LifoQueue
 from typing import Dict, List, Generic, TypeVar, Optional, Generator
 
-from yk_gmd_blender.yk_gmd.v2.abstract.gmd_node import GMDBone, GMDSkinnedObject, GMDUnskinnedObject, GMDNode
+from yk_gmd_blender.yk_gmd.v2.abstract.nodes.gmd_bone import GMDBone
+from yk_gmd_blender.yk_gmd.v2.abstract.nodes.gmd_node import GMDNode
+from yk_gmd_blender.yk_gmd.v2.abstract.nodes.gmd_object import GMDSkinnedObject, GMDUnskinnedObject
 
 TNode = TypeVar('TNode', bound=GMDNode)
