+++ yk_gmd_blender/yk_gmd/v2/abstract/gmd_shader.py
@@ -1,1 +1,1 @@
-from dataclasses import dataclassfrom enum import Enumfrom typing import Optional, Tuple, Listfrom mathutils import Vectorfrom yk_gmd_blender.structurelib.base import StructureUnpacker, BasePrimitive, FixedSizeArrayUnpacker, ValueAdaptor, \    BaseUnpackerfrom yk_gmd_blender.structurelib.primitives import c_float32, c_float16, c_unorm8, c_uint8, RangeConverterPrimitivefrom yk_gmd_blender.yk_gmd.v2.structure.common.vector import Vec3Unpacker, Vec4Unpacker, Vec3Unpacker_of, \    Vec4Unpacker_of# TODO: Rename to HLSL-style (ubyte4, float2, float3, float4, half2 etc.)class VecStorage(Enum):    Vec4Fixed = 1    Vec2Full = 2    Vec3Full = 3    Vec4Full = 4    Vec2Half = 5    Vec3Half = 6    Vec4Half = 7class FixedConvertMethod(Enum):    To0_1 = 0    ToMinus1_1 = 1    ToByte = 2def list2_to_tuple2(list2: List):    return (list2[0], list2[1])def make_vector_unpacker(vec_type: VecStorage, byte_convert: Optional[FixedConvertMethod] = None) -> BaseUnpacker:    if vec_type == VecStorage.Vec2Full:        return ValueAdaptor(tuple, base_unpacker=FixedSizeArrayUnpacker(c_float32, 2), forwards=list2_to_tuple2, backwards=None)    elif vec_type == VecStorage.Vec3Full:        return Vec3Unpacker_of(c_float32)    elif vec_type == VecStorage.Vec4Full:        return Vec4Unpacker_of(c_float32)    elif vec_type == VecStorage.Vec2Half:        return ValueAdaptor(tuple, base_unpacker=FixedSizeArrayUnpacker(c_float16, 2), forwards=list2_to_tuple2, backwards=None)    elif vec_type == VecStorage.Vec3Half:        return Vec3Unpacker_of(c_float16)    elif vec_type == VecStorage.Vec4Half:        return Vec4Unpacker_of(c_float16)    elif vec_type == VecStorage.Vec4Fixed:        if not byte_convert:            print(f"Field had vec_type = Vec4Fixed but gave no byte_convert, assuming ToByte")            byte_convert = FixedConvertMethod.ToByte        if byte_convert == FixedConvertMethod.To0_1:            return Vec4Unpacker_of(c_unorm8)        elif byte_convert == FixedConvertMethod.ToMinus1_1:            return Vec4Unpacker_of(RangeConverterPrimitive(base_unpack=c_uint8, from_range=(0, 255), to_range=(-1.0, 1.0)))        elif byte_convert == FixedConvertMethod.ToByte:            return Vec4Unpacker_of(c_uint8)    raise TypeError(f"Couldn't generate vector unpacker for {vec_type}/{byte_convert}")@dataclassclass BoneWeight:    bone: int    weight: floatBoneWeight4 = Tuple[BoneWeight, BoneWeight, BoneWeight, BoneWeight]@dataclass(repr=False)class GMDVertexBuffer:    layout: 'GMDVertexBufferLayout'    pos: List[Vector]    bone_weights: Optional[List[BoneWeight4]]    normal: Optional[List[Vector]]    tangent: Optional[List[Vector]]    unk: Optional[List[Vector]]    col0: Optional[List[Vector]]    col1: Optional[List[Vector]]    uvs: List[List[Vector]]    @staticmethod    def build_empty(layout: 'GMDVertexBufferLayout', count: int):        alloc_for_type = lambda field_type: [None] * count if field_type else None        return GMDVertexBuffer(            layout=layout,            # TODO: How to deal with "incomplete" vertex buffers?            pos = [None] * count,            bone_weights = alloc_for_type(layout.weights_type),            normal = alloc_for_type(layout.normal_type),            tangent = alloc_for_type(layout.tangent_type),            unk = alloc_for_type(layout.unk_type),            col0 = alloc_for_type(layout.col0_type),            col1 = alloc_for_type(layout.col1_type),            uvs = [                [None] * count                for t in layout.uv_storages            ],        )    def vertex_count(self):        return len(self.pos)    def __getitem__(self, item) -> 'GMDVertexBuffer':        if not isinstance(item, slice):            raise IndexError(f"GMDVertexBuffer __getitem__ got {item} but expected slice")        return GMDVertexBuffer(            layout=self.layout,            pos=self.pos[item],            bone_weights=self.bone_weights[item],            normal=self.normal[item],            tangent=self.tangent[item],            unk=self.unk[item],            col0=self.col0[item],            col1=self.col1[item],            uvs=[                uv[item]                for uv in self.uvs            ],        )# VertexBufferLayouts are external dependencies (shaders have a fixed layout, which we can't control) so they are frozen@dataclass(frozen=True, init=True)class GMDVertexBufferLayout:    pos_type: BaseUnpacker    weights_type: Optional[BaseUnpacker]    bones_type: Optional[BaseUnpacker]    normal_type: Optional[BaseUnpacker]    tangent_type: Optional[BaseUnpacker]    unk_type: Optional[BaseUnpacker]    col0_type: Optional[BaseUnpacker]    col1_type: Optional[BaseUnpacker]    uv_storages: List[BaseUnpacker]    def __post_init__(self):        if not self.pos_type:            raise TypeError(f"VertexBufferLayout claims to not have vertex positions!")        if bool(self.weights_type) != bool(self.bones_type):            raise TypeError(f"VertexBufferLayout has weights/bones mismatch: weights {bool(self.weights_type)} bones {bool(self.bones_type)}")    @staticmethod    def make_vertex_buffer_layout(                                 pos_storage: VecStorage,                                 weights_storage: Optional[VecStorage],                                 bones_storage: Optional[VecStorage],                                 normal_storage: Optional[VecStorage],                                 tangent_storage: Optional[VecStorage],                                 unk_storage: Optional[VecStorage],                                 col0_storage: Optional[VecStorage],                                 col1_storage: Optional[VecStorage],                                 uv_storages: List[VecStorage],                                 ):        return GMDVertexBufferLayout(            pos_type=make_vector_unpacker(pos_storage),            weights_type=make_vector_unpacker(weights_storage, FixedConvertMethod.To0_1) if weights_storage else None,            bones_type=make_vector_unpacker(bones_storage, FixedConvertMethod.ToByte) if bones_storage else None,            normal_type=make_vector_unpacker(normal_storage, FixedConvertMethod.ToMinus1_1) if normal_storage else None,            tangent_type=make_vector_unpacker(tangent_storage, FixedConvertMethod.ToMinus1_1) if tangent_storage else None,            unk_type=make_vector_unpacker(unk_storage) if unk_storage else None,            col0_type=make_vector_unpacker(col0_storage, FixedConvertMethod.To0_1) if col0_storage else None,            col1_type=make_vector_unpacker(col1_storage, FixedConvertMethod.To0_1) if col1_storage else None,            uv_storages=[make_vector_unpacker(s, FixedConvertMethod.ToMinus1_1) for s in uv_storages],        )    def bytes_per_vertex(self) -> int:        size = 0        size += self.pos_type.sizeof()        size += self.weights_type.sizeof()        size += self.bones_type.sizeof()        if self.normal_type:            size += self.normal_type.sizeof()        if self.tangent_type:            size += self.tangent_type.sizeof()        if self.unk_type:            size += self.unk_type.sizeof()        if self.col0_type:            size += self.col0_type.sizeof()        if self.col1_type:            size += self.col1_type.sizeof()        for uv_type in self.uv_storages:            size += uv_type.sizeof()        return size    def unpack_from(self, big_endian: bool, vertex_count: int, data: bytes, offset: int) -> Tuple[GMDVertexBuffer, int]:        vertices: GMDVertexBuffer = GMDVertexBuffer.build_empty(self, vertex_count)        for i in range(vertex_count):            # unpack() returns the unpacked value and offset + size, so incrementing offset is done in one line            vertices.pos[i], offset = self.pos_type.unpack(big_endian, data, offset)            if self.weights_type:                weights, offset = self.weights_type.unpack(big_endian, data, offset)                bones, offset = self.bones_type.unpack(big_endian, data, offset)                vertices.bone_weights[i] = (                    BoneWeight(bone=bones[0], weight=weights[0]),                    BoneWeight(bone=bones[1], weight=weights[1]),                    BoneWeight(bone=bones[2], weight=weights[2]),                    BoneWeight(bone=bones[3], weight=weights[3]),                )            if self.normal_type:                vertices.normal[i], offset = self.normal_type.unpack(big_endian, data, offset)            if self.tangent_type:                vertices.tangent[i], offset = self.tangent_type.unpack(big_endian, data, offset)            if self.unk_type:                vertices.unk[i], offset = self.unk_type.unpack(big_endian, data, offset)            if self.col0_type:                vertices.col0[i], offset = self.col0_type.unpack(big_endian, data, offset)            if self.col1_type:                vertices.col1[i], offset = self.col1_type.unpack(big_endian, data, offset)            for uv_idx, uv_storage in enumerate(self.uv_storages):                vertices.uvs[uv_idx][i], offset = uv_storage.unpack(big_endian, data, offset)        return vertices, offset    # def check_verts(self, vertices: List[GMDVertex]):    #     for vertex in vertices:    #         if not all(vertex.pos, vertex.bone_weights):    def pack_into(self, big_endian: bool, vertices: GMDVertexBuffer, append_to: bytearray):        # TODO: Validate that all ranges exist for what we want        # TODO: Fill in default data if stuff is missing?        for i in range(vertices.vertex_count()):            self.pos_type.pack(big_endian, vertices.pos[i], append_to=append_to)            if self.weights_type:                self.weights_type.pack(big_endian, [b.weight for b in vertices.bone_weights[i]], append_to=append_to)                self.bones_type.pack(big_endian, [b.bone for b in vertices.bone_weights[i]], append_to=append_to)            if self.normal_type:                self.normal_type.pack(big_endian, vertices.normal[i], append_to=append_to)            if self.tangent_type:                self.tangent_type.pack(big_endian, vertices.tangent[i], append_to=append_to)            if self.unk_type:                self.unk_type.pack(big_endian, vertices.unk[i], append_to=append_to)            if self.col0_type:                self.col0_type.pack(big_endian, vertices.col0[i], append_to=append_to)            if self.col1_type:                self.col1_type.pack(big_endian, vertices.col1[i], append_to=append_to)            for uv_data, uv_packer in zip(vertices.uvs, self.uv_storages):                uv_packer.pack(big_endian, uv_data[i], append_to=append_to)# Shaders are external dependencies, so they are frozen. You can't change the name of a shader, for example.@dataclass(frozen=True)class GMDShader:    name: str    vertex_buffer_layout: GMDVertexBufferLayout
+from dataclasses import dataclassfrom enum import Enumfrom typing import Optional, Tuple, Listfrom mathutils import Vectorfrom yk_gmd_blender.structurelib.base import StructureUnpacker, BasePrimitive, FixedSizeArrayUnpacker, ValueAdaptor, \    BaseUnpackerfrom yk_gmd_blender.structurelib.primitives import c_float32, c_float16, c_unorm8, c_uint8, RangeConverterPrimitivefrom yk_gmd_blender.yk_gmd.v2.structure.common.vector import Vec3Unpacker, Vec4Unpacker, Vec3Unpacker_of, \    Vec4Unpacker_of# TODO: Rename to HLSL-style (ubyte4, float2, float3, float4, half2 etc.)class VecStorage(Enum):    Vec4Fixed = 1    Vec2Full = 2    Vec3Full = 3    Vec4Full = 4    Vec2Half = 5    Vec3Half = 6    Vec4Half = 7class FixedConvertMethod(Enum):    To0_1 = 0    ToMinus1_1 = 1    ToByte = 2def list2_to_tuple2(list2: List):    return (list2[0], list2[1])def make_vector_unpacker(vec_type: VecStorage, byte_convert: Optional[FixedConvertMethod] = None) -> BaseUnpacker:    if vec_type == VecStorage.Vec2Full:        return ValueAdaptor(tuple, base_unpacker=FixedSizeArrayUnpacker(c_float32, 2), forwards=list2_to_tuple2, backwards=None)    elif vec_type == VecStorage.Vec3Full:        return Vec3Unpacker_of(c_float32)    elif vec_type == VecStorage.Vec4Full:        return Vec4Unpacker_of(c_float32)    elif vec_type == VecStorage.Vec2Half:        return ValueAdaptor(tuple, base_unpacker=FixedSizeArrayUnpacker(c_float16, 2), forwards=list2_to_tuple2, backwards=None)    elif vec_type == VecStorage.Vec3Half:        return Vec3Unpacker_of(c_float16)    elif vec_type == VecStorage.Vec4Half:        return Vec4Unpacker_of(c_float16)    elif vec_type == VecStorage.Vec4Fixed:        if not byte_convert:            print(f"Field had vec_type = Vec4Fixed but gave no byte_convert, assuming ToByte")            byte_convert = FixedConvertMethod.ToByte        if byte_convert == FixedConvertMethod.To0_1:            return Vec4Unpacker_of(c_unorm8)        elif byte_convert == FixedConvertMethod.ToMinus1_1:            return Vec4Unpacker_of(RangeConverterPrimitive(base_unpack=c_uint8, from_range=(0, 255), to_range=(-1.0, 1.0)))        elif byte_convert == FixedConvertMethod.ToByte:            return Vec4Unpacker_of(c_uint8)    raise TypeError(f"Couldn't generate vector unpacker for {vec_type}/{byte_convert}")@dataclassclass BoneWeight:    bone: int    weight: floatBoneWeight4 = Tuple[BoneWeight, BoneWeight, BoneWeight, BoneWeight]@dataclass(repr=False)class GMDVertexBuffer:    layout: 'GMDVertexBufferLayout'    pos: List[Vector]    bone_weights: Optional[List[BoneWeight4]]    normal: Optional[List[Vector]]    tangent: Optional[List[Vector]]    unk: Optional[List[Vector]]    col0: Optional[List[Vector]]    col1: Optional[List[Vector]]    uvs: List[List[Vector]]    @staticmethod    def build_empty(layout: 'GMDVertexBufferLayout', count: int):        alloc_for_type = lambda field_type: [None] * count if field_type else None        return GMDVertexBuffer(            layout=layout,            # TODO: How to deal with "incomplete" vertex buffers?            pos = [None] * count,            bone_weights = alloc_for_type(layout.weights_type),            normal = alloc_for_type(layout.normal_type),            tangent = alloc_for_type(layout.tangent_type),            unk = alloc_for_type(layout.unk_type),            col0 = alloc_for_type(layout.col0_type),            col1 = alloc_for_type(layout.col1_type),            uvs = [                [None] * count                for t in layout.uv_storages            ],        )    def vertex_count(self):        return len(self.pos)    def __getitem__(self, item) -> 'GMDVertexBuffer':        if not isinstance(item, slice):            raise IndexError(f"GMDVertexBuffer __getitem__ got {item} but expected slice")        return GMDVertexBuffer(            layout=self.layout,            pos=self.pos[item],            bone_weights=self.bone_weights[item],            normal=self.normal[item],            tangent=self.tangent[item],            unk=self.unk[item],            col0=self.col0[item],            col1=self.col1[item],            uvs=[                uv[item]                for uv in self.uvs            ],        )# VertexBufferLayouts are external dependencies (shaders have a fixed layout, which we can't control) so they are frozen@dataclass(frozen=True, init=True)class GMDVertexBufferLayout:    pos_type: BaseUnpacker    weights_type: Optional[BaseUnpacker]    bones_type: Optional[BaseUnpacker]    normal_type: Optional[BaseUnpacker]    tangent_type: Optional[BaseUnpacker]    unk_type: Optional[BaseUnpacker]    col0_type: Optional[BaseUnpacker]    col1_type: Optional[BaseUnpacker]    uv_storages: List[BaseUnpacker]    def __post_init__(self):        if not self.pos_type:            raise TypeError(f"VertexBufferLayout claims to not have vertex positions!")        if bool(self.weights_type) != bool(self.bones_type):            raise TypeError(f"VertexBufferLayout has weights/bones mismatch: weights {bool(self.weights_type)} bones {bool(self.bones_type)}")    @staticmethod    def make_vertex_buffer_layout(                                 pos_storage: VecStorage,                                 weights_storage: Optional[VecStorage],                                 bones_storage: Optional[VecStorage],                                 normal_storage: Optional[VecStorage],                                 tangent_storage: Optional[VecStorage],                                 unk_storage: Optional[VecStorage],                                 col0_storage: Optional[VecStorage],                                 col1_storage: Optional[VecStorage],                                 uv_storages: List[VecStorage],                                 ):        return GMDVertexBufferLayout(            pos_type=make_vector_unpacker(pos_storage),            weights_type=make_vector_unpacker(weights_storage, FixedConvertMethod.To0_1) if weights_storage else None,            bones_type=make_vector_unpacker(bones_storage, FixedConvertMethod.ToByte) if bones_storage else None,            normal_type=make_vector_unpacker(normal_storage, FixedConvertMethod.ToMinus1_1) if normal_storage else None,            tangent_type=make_vector_unpacker(tangent_storage, FixedConvertMethod.ToMinus1_1) if tangent_storage else None,            unk_type=make_vector_unpacker(unk_storage) if unk_storage else None,            col0_type=make_vector_unpacker(col0_storage, FixedConvertMethod.To0_1) if col0_storage else None,            col1_type=make_vector_unpacker(col1_storage, FixedConvertMethod.To0_1) if col1_storage else None,            uv_storages=[make_vector_unpacker(s, FixedConvertMethod.ToMinus1_1) for s in uv_storages],        )    def bytes_per_vertex(self) -> int:        size = 0        size += self.pos_type.sizeof()        if self.weights_type:            size += self.weights_type.sizeof()            size += self.bones_type.sizeof()        if self.normal_type:            size += self.normal_type.sizeof()        if self.tangent_type:            size += self.tangent_type.sizeof()        if self.unk_type:            size += self.unk_type.sizeof()        if self.col0_type:            size += self.col0_type.sizeof()        if self.col1_type:            size += self.col1_type.sizeof()        for uv_type in self.uv_storages:            size += uv_type.sizeof()        return size    def unpack_from(self, big_endian: bool, vertex_count: int, data: bytes, offset: int) -> Tuple[GMDVertexBuffer, int]:        vertices: GMDVertexBuffer = GMDVertexBuffer.build_empty(self, vertex_count)        for i in range(vertex_count):            # unpack() returns the unpacked value and offset + size, so incrementing offset is done in one line            vertices.pos[i], offset = self.pos_type.unpack(big_endian, data, offset)            if self.weights_type:                weights, offset = self.weights_type.unpack(big_endian, data, offset)                bones, offset = self.bones_type.unpack(big_endian, data, offset)                vertices.bone_weights[i] = (                    BoneWeight(bone=bones[0], weight=weights[0]),                    BoneWeight(bone=bones[1], weight=weights[1]),                    BoneWeight(bone=bones[2], weight=weights[2]),                    BoneWeight(bone=bones[3], weight=weights[3]),                )            if self.normal_type:                vertices.normal[i], offset = self.normal_type.unpack(big_endian, data, offset)            if self.tangent_type:                vertices.tangent[i], offset = self.tangent_type.unpack(big_endian, data, offset)            if self.unk_type:                vertices.unk[i], offset = self.unk_type.unpack(big_endian, data, offset)            if self.col0_type:                vertices.col0[i], offset = self.col0_type.unpack(big_endian, data, offset)            if self.col1_type:                vertices.col1[i], offset = self.col1_type.unpack(big_endian, data, offset)            for uv_idx, uv_storage in enumerate(self.uv_storages):                vertices.uvs[uv_idx][i], offset = uv_storage.unpack(big_endian, data, offset)        return vertices, offset    # def check_verts(self, vertices: List[GMDVertex]):    #     for vertex in vertices:    #         if not all(vertex.pos, vertex.bone_weights):    def pack_into(self, big_endian: bool, vertices: GMDVertexBuffer, append_to: bytearray):        # TODO: Validate that all ranges exist for what we want        # TODO: Fill in default data if stuff is missing?        for i in range(vertices.vertex_count()):            self.pos_type.pack(big_endian, vertices.pos[i], append_to=append_to)            if self.weights_type:                self.weights_type.pack(big_endian, [b.weight for b in vertices.bone_weights[i]], append_to=append_to)                self.bones_type.pack(big_endian, [b.bone for b in vertices.bone_weights[i]], append_to=append_to)            if self.normal_type:                self.normal_type.pack(big_endian, vertices.normal[i], append_to=append_to)            if self.tangent_type:                self.tangent_type.pack(big_endian, vertices.tangent[i], append_to=append_to)            if self.unk_type:                self.unk_type.pack(big_endian, vertices.unk[i], append_to=append_to)            if self.col0_type:                self.col0_type.pack(big_endian, vertices.col0[i], append_to=append_to)            if self.col1_type:                self.col1_type.pack(big_endian, vertices.col1[i], append_to=append_to)            for uv_data, uv_packer in zip(vertices.uvs, self.uv_storages):                uv_packer.pack(big_endian, uv_data[i], append_to=append_to)# Shaders are external dependencies, so they are frozen. You can't change the name of a shader, for example.@dataclass(frozen=True)class GMDShader:    name: str    vertex_buffer_layout: GMDVertexBufferLayout
